feat: enhance WebView cookie management and state handling
